zorko-designer-53 index.ts:11-13: Move middleware to index.ts extract... - move store creation to main index
