+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {render} from 'react-dom';
 import './index.css';
 import {Provider} from 'react-redux';
-import store from './store';
 import {repositoriesLoadInitial} from './features/repositories/asyncActions';
 import {AppLogger} from './app/AppLogger';
 import vegaLiteSchema from './defaultVegaLiteSchema.json';
@@ -10,6 +9,13 @@ import {vegaLiteSchemaReadSuccess} from './features/vegaLiteSchema';
 import {chooseSpecFlow} from './features/chooseSpecFlow/asyncActions';
 import {AppShell} from './app/AppShell';
 import {HashRouter} from 'react-router-dom';
+import {createStore} from './store/createStore';
+import {reduxLoggerMiddleware} from './logger/reduxLoggerMiddleware';
+import {logrockMiddleware} from './logger/logrockMiddleware';
+import firebase from 'firebase';
+import {firebaseConfig} from '../firebase.config';
+import {ZorkoDesignerAnalyticFacade} from './analytic/ZorkoDesignerAnalyticFacade';
+import {zorkoDesignerAnalyticMiddleware} from './features/analytic/zorkoDesignerAnalyticMiddleware';
 
 declare global {
   interface Window {
@@ -17,10 +23,24 @@ declare global {
   }
 }
 
+// Define uniq session for logging proposes
 window.sessionID = `sessionid-${Math.random()
   .toString(36)
   .substr(3, 9)}`;
 
+const middleware = [reduxLoggerMiddleware, logrockMiddleware];
+
+if (process.env.NODE_ENV === 'production') {
+  const app = firebase.initializeApp(firebaseConfig);
+  const analytics = app.analytics();
+  const analyticFacade = new ZorkoDesignerAnalyticFacade(analytics);
+
+  middleware.push(zorkoDesignerAnalyticMiddleware(analyticFacade));
+}
+
+const store = createStore(middleware);
+
+// Fire few action so user can see some pre-selected visualization
 store.dispatch(vegaLiteSchemaReadSuccess(vegaLiteSchema));
 store.dispatch(repositoriesLoadInitial());
 store.dispatch(chooseSpecFlow('bar'));
